+++ src/main/java/com/jajaja/domain/order/dto/response/TossPayments/CardDto.java
@@ -1,4 +1,4 @@
-package com.jajaja.domain.order.dto.response;
+package com.jajaja.domain.order.dto.response.TossPayments;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
